divide functions to seperate files
