--- crawler.py
@@ -1,14 +1,0 @@
-def foo():
-    pass
-
-
-def bar():
-    pass
-
-
-def main():
-    pass
-
-
-if __name__ == "__main__":
-    main()
